requirements.txt
langchain-openai==0.0.5
langchain-prompts==0.0.5
langchain-elasticsearch==0.0.5
langchain-community==0.0.5
langchain-schema==0.0.5
openai==0.27.0
elasticsearch==8.0.0
